Update query format
